GamePlay: Win, Lost, Tie popup

+++ server/services/player_services.js
@@ -22,8 +22,12 @@ const get_player_init_info = async (game_sockets, socket_id, game_id) => {
   return newPlayer;
 };
 
-const remove_player = async (socket_id) => {
-  await Player.deleteOne({ socket_id: socket_id }).exec();
+const remove_player = async (io, socket_id) => {
+  const playerData = await Player.findOne({ socket_id: socket_id });
+  if (playerData) {
+    await io.to(playerData.game_id.toString()).emit("game_ready", false);
+    await Player.deleteOne({ socket_id: socket_id }).exec();
+  }
 };
 
 module.exports.remove_player = remove_player;
